src/auth/requirements.txt
astroid
click
dill
flake8
Flask
Flask-MySQLdb
isort
itsdangerous
jedi
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
mysqlclient
parso
platformdirs
pycodestyle
pyflakes
PyJWT
pylint
toml
tomli
tomlkit
Werkzeug
wrapt
